Introduce AbstractCreateIssue class
